refactor(coral): Rename table-filters -> filters, useTableFiltersValues -> useFiltersValues
