+++ coral/src/app/features/components/filters/TeamFilter.test.tsx
@@ -1,7 +1,7 @@
 import { cleanup, screen, waitFor } from "@testing-library/react";
 import { waitForElementToBeRemoved } from "@testing-library/react/pure";
 import userEvent from "@testing-library/user-event";
-import TeamFilter from "src/app/features/components/table-filters/TeamFilter";
+import TeamFilter from "src/app/features/components/filters/TeamFilter";
 import { getTeams } from "src/domain/team/team-api";
 import { customRender } from "src/services/test-utils/render-with-wrappers";
